fderiv: fix newdata generation if not specified
Creating the newdata data frame now works with factors and an
arbitrary data frame length. Before, it only worked with factors if
the data frame length happened to be a multiple of the n argument to
fderiv (the call to data.frame, from cbind, would then lead to
recycling).

To see an example of the bug that this commit fixes, consider the
following:

  library(mgcv)
  library(tsgam)
  dat <- gamSim(4, n = 401, dist = "normal", scale = 2)
  mod <- gam(y ~ s(x0) + s(x1) + fac, data = dat, method = "REML")
  fd <- fderiv(mod)

Before this commit, the example gave an error:

  Error in data.frame(..., check.names = FALSE) :
    arguments imply differing number of rows: 401, 200