fderiv: fix factor tracking if newdata specified
In order for the logical vector that tracks factor columns to be
valid when selecting columns, the response variable should be
removed. This was already done when the newdata data frame was not
specified by the user, but it was not done when the newdata frame
was specified by the user.

To see an example of the bug that this commit fixes, consider the
following:

  library(mgcv)
  library(tsgam)
  dat <- gamSim(4, n = 400, dist = "normal", scale = 2)
  mod <- gam(y ~ s(x0) + s(x1) + fac, data = dat, method = "REML")
  newd <- dat[1,]
  fd <- fderiv(mod, newdata = newd)

Before this commit, the example gave an error:

  Error in `[.data.frame`(newdata2, , !ff, drop = FALSE) :
    undefined columns selected